docs: move schema metadata to customization
