+++ packages/docs/package.json
@@ -41,7 +41,6 @@
     "@fn-sphere/core": "workspace:*",
     "@fn-sphere/filter": "workspace:*",
     "@fn-sphere/theme-mui-material": "workspace:*",
-    "@mui/icons-material": "^9.0.0",
     "@mui/material": "^9.2.0",
     "zod": "^4.4.3"
   }
