+++ packages/docs/src/content/docs/customization/localization.mdx
@@ -101,7 +101,7 @@ export default function AdvancedFilter() {
 
 You can customize the field name and filter name by providing the `mapFieldName` and `mapFilterName` functions.
 
-Field descriptions are Zod schema metadata. See [Using Schema Metadata](/fn-sphere/guides/best-practices/#using-schema-metadata) to learn how to define typed metadata keys and read them in custom consumers.
+Field descriptions are Zod schema metadata. See [Schema Metadata](/fn-sphere/customization/schema-metadata/) to learn how to define typed metadata keys and read them in custom consumers.
 
 ```tsx live "mapFieldName" "mapFilterName"
 import {
