+++ packages/docs/src/content/docs/customization/schema-metadata.mdx
@@ -1,0 +1,80 @@
+---
+title: Schema Metadata
+description: Attach typed metadata to Zod schemas and consume it in custom filter UI
+---
+
+import SourceCodeCard from "~/components/source-code-card.astro";
+
+Zod schema metadata lets your application attach configuration to a field and read it wherever that schema is available. Filter Sphere exposes each field's original schema, so custom UI can consume metadata without changing filter rules or Filter Sphere's public APIs.
+
+The data flow is:
+
+`schema field .meta()` → `z.globalRegistry` → `FilterField.fieldSchema` → custom consumer
+
+## Define Typed Metadata
+
+Application-defined metadata keys can be type-checked by augmenting Zod's `GlobalMeta` interface. Keep the values independent from a specific rendering library. In this example, `icon` stores a semantic token rather than a React component:
+
+```ts
+import { z } from "zod";
+
+type IconName = "user" | "email" | "number" | "calendar";
+
+declare module "zod" {
+  interface GlobalMeta {
+    icon?: IconName;
+  }
+}
+
+export const schema = z.object({
+  name: z.string().meta({ description: "Name", icon: "user" }),
+  email: z.string().meta({ description: "Email", icon: "email" }),
+  age: z.number().meta({ description: "Age", icon: "number" }),
+  createdAt: z.date().meta({ description: "Created At", icon: "calendar" }),
+  notes: z.string().meta({ description: "Notes" }),
+});
+```
+
+`description` is built-in Zod metadata. Keys such as `icon` are application conventions and are not part of the Filter Sphere API.
+
+## Read Metadata from a Filter Field
+
+Filter Sphere exposes the registered schema instance as `FilterField.fieldSchema`. Read it directly from Zod's global registry; no schema cast is needed:
+
+```tsx
+import { useFilterSelect, type FilterField } from "@fn-sphere/filter";
+import { z } from "zod";
+
+const getFieldIcon = (field: FilterField | undefined) =>
+  field ? z.globalRegistry.get(field.fieldSchema)?.icon : undefined;
+
+function FieldSelectOptions({ rule }) {
+  const { selectedField, fieldOptions } = useFilterSelect(rule);
+  const selectedIcon = getFieldIcon(selectedField);
+  const options = fieldOptions.map((option) => ({
+    ...option,
+    icon: getFieldIcon(option.value),
+  }));
+
+  // Render selectedIcon and options with any UI technology.
+}
+```
+
+The same approach works for labels, analytics identifiers, permissions, formatting hints, or other configuration owned by your application.
+
+## Live Example
+
+This example uses metadata to add different glyphs to a custom `FieldSelect`. It overrides only the `FieldSelect` template and reuses Filter Sphere's native `<select>`, so it has no UI-library dependency. The `notes` field intentionally has no `icon` and demonstrates the plain-label fallback.
+
+import { SchemaMetadataExample } from "~/components/examples/schema-metadata.tsx";
+
+<div className="not-content">
+  <SchemaMetadataExample client:load />
+</div>
+<SourceCodeCard path="~/components/examples/schema-metadata.tsx" />
+
+## Caveats
+
+- Keep `.meta()` at the end of a schema construction chain. It returns the schema instance registered with that metadata, so the returned instance must be the one placed in the object shape.
+- Store serializable, application-owned values such as semantic tokens. Resolve those tokens to components, glyphs, or other presentation inside the consumer.
+- Metadata remains associated with the schema. Filter Sphere does not copy it into `SingleFilter` rules, and serializing a rule does not serialize schema metadata.
