+++ packages/docs/src/content/docs/customization/theme.mdx
@@ -235,5 +235,7 @@ import { ThemeAnatomy } from "~/components/theme-anatomy.tsx";
       - `FilterDataInput`: The input field for the rule.
     - `RuleJoiner`: Used to render the joiner between the rules, it will render nothing by default.
 
+    See [Schema Metadata](/fn-sphere/customization/schema-metadata/) for an example that reads application-owned field metadata from a custom `FieldSelect`.
+
   </TabItem>
 </Tabs>
